pyproject.toml
[project]
name = "openai-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "misbahkhan99", email = "misbahmubashara98@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.1",
    "openai-agents>=0.0.4",
    "python-dotenv>=1.0.1",
]

[project.scripts]
openai-agent-project = "openai_agent_project:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
